Implement core functionality for mini-env, including state management and API endpoints for episode handling, step execution, and verification. Update reward type from f64 to f32 in response structures for consistency.

+++ src/types.rs
@@ -40,7 +40,7 @@ pub struct StepRequest {
 #[derive(Debug, Serialize, Deserialize)]
 pub struct StepResponse {
     pub observation: String,
-    pub reward: f64,
+    pub reward: f32,
     pub done: bool,
     pub info: StepInfo,
 }
@@ -59,7 +59,7 @@ pub struct VerifyRequest {
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct VerifyResponse {
-    pub reward: f64,
+    pub reward: f32,
     pub correctness: f64,
     pub test_integrity: bool,
     pub tests_passed: u32,
